requirements.txt
anyio==4.4.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
google-ai-generativelanguage
google-api-core==2.19.0
google-auth==2.29.0
google-generativeai==0.5.4
googleapis-common-protos==1.63.0
grpcio==1.64.0
grpcio-status==1.60.0
h11==0.14.0
httpcore==1.0.5
idna==3.7
proto-plus==1.23.0
protobuf
pyasn1==0.6.0
pyasn1-modules==0.4.0
python-dotenv==1.0.1
python-telegram-bot==21.2
requests==2.32.2
rsa==4.9
sniffio==1.3.1
tqdm==4.66.4
urllib3==2.2.1
Pillow==10.3.0